requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pyyaml==6.0.2
